requirements.txt
pillow
tqdm
python-dotenv
together
python-opencv